requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
gigachat==0.1.8
requests==2.31.0
pytest==7.4.3